feat: send unmodified first frame so that the video stream doesn't show a gray image
Multiple synchronous functions were being called that took time for
`transform` to return, and this meant that while the event loop was
locked up in that sync code, it couldn't render frames.

So, before processing the first frame, enqueue that unprocessed frame
and then wait a few milliseconds for the frame to get written to the
screen before continuing with all the heavy sync processing, which means
that effectively instead of showing a gray image, it shows the first
frame instead which looks visually less jarring.